GoogleCloudRun/requirements.txt
flask==2.0.3
werkzeug==2.0.3
gunicorn==20.1.0
openai==1.59.3
python-dotenv==1.0.0
ffmpeg-python==0.2.0
geopy==2.3.0
spacy==3.7.2
numpy==1.24.2
contractions==0.1.73
certifi==2023.5.7
whisper>=1.0.0
moviepy==1.0.3
pytube==12.1.2
scikit-learn==1.2.2
google-cloud-storage==2.9.0
spacy-transformers>=1.2.2
spacy-legacy>=3.0.12
spacy-loggers>=1.0.5
auto-labeling-pipeline>=0.1.23
spacy-alignments>=0.9.1
spacy-lookups-data>=1.0.5
pathlib>=1.0.1
safetensors>=0.4.0
torch>=2.0.0
sentence-transformers==3.4.1
huggingface-hub>=0.19.0
transformers>=4.34.0